# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"HealthResponse": {
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": [
					"ok"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentCurrency": {
				"type": "string",
				"enum": [
					"COP",
					"USD"
				]
			},
			"TimelineChartItem": {
				"properties": {
					"period": {
						"type": "string"
					},
					"currency": {
						"$ref": "#/components/schemas/PaymentCurrency"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"period",
					"currency",
					"amount",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimelineGranularity": {
				"type": "string",
				"enum": [
					"day",
					"week",
					"month"
				]
			},
			"PaymentStatus": {
				"type": "string",
				"enum": [
					"completed",
					"failed",
					"refunded"
				]
			},
			"CountMetric": {
				"properties": {
					"currency": {
						"$ref": "#/components/schemas/PaymentCurrency"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"currency",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MoneyMetric": {
				"properties": {
					"currency": {
						"$ref": "#/components/schemas/PaymentCurrency"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"currency",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentsSummary": {
				"properties": {
					"totalPayments": {
						"type": "number",
						"format": "double"
					},
					"totalRefunds": {
						"type": "number",
						"format": "double"
					},
					"paymentsByCurrency": {
						"items": {
							"$ref": "#/components/schemas/CountMetric"
						},
						"type": "array"
					},
					"refundsByCurrency": {
						"items": {
							"$ref": "#/components/schemas/CountMetric"
						},
						"type": "array"
					},
					"completedRevenueByCurrency": {
						"items": {
							"$ref": "#/components/schemas/MoneyMetric"
						},
						"type": "array"
					},
					"averageTicketByCurrency": {
						"items": {
							"$ref": "#/components/schemas/MoneyMetric"
						},
						"type": "array"
					},
					"lastUpdated": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"totalPayments",
					"totalRefunds",
					"paymentsByCurrency",
					"refundsByCurrency",
					"completedRevenueByCurrency",
					"averageTicketByCurrency",
					"lastUpdated"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StatusChartItem": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/PaymentStatus"
					},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Payment": {
				"properties": {
					"id_pago": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"nullable": true
					},
					"nombre": {
						"type": "string",
						"nullable": true
					},
					"curso": {
						"type": "string"
					},
					"importe": {
						"type": "number",
						"format": "double"
					},
					"moneda": {
						"$ref": "#/components/schemas/PaymentCurrency"
					},
					"estado": {
						"$ref": "#/components/schemas/PaymentStatus"
					},
					"fecha": {
						"type": "string"
					},
					"refunded_at": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id_pago",
					"curso",
					"importe",
					"moneda",
					"estado",
					"fecha"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentSortBy": {
				"type": "string",
				"enum": [
					"id_pago",
					"email",
					"nombre",
					"curso",
					"importe",
					"moneda",
					"estado",
					"fecha",
					"refunded_at"
				]
			},
			"SortOrder": {
				"type": "string",
				"enum": [
					"asc",
					"desc"
				]
			},
			"PaginatedPaymentsResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Payment"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"pageSize": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"pageSize",
							"page"
						],
						"type": "object"
					},
					"sort": {
						"properties": {
							"sortOrder": {
								"$ref": "#/components/schemas/SortOrder"
							},
							"sortBy": {
								"$ref": "#/components/schemas/PaymentSortBy"
							}
						},
						"required": [
							"sortOrder",
							"sortBy"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination",
					"sort"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "FormaPro Payments Dashboard API",
		"version": "1.0.0",
		"description": "API de solo lectura para consultar pagos, KPIs, gráficos y exportación CSV desde Supabase operations.payments."
	},
	"paths": {
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		},
		"/payments/charts/revenue-timeline": {
			"get": {
				"operationId": "GetPaymentsTimeline",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TimelineChartItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Payments Charts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "granularity",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TimelineGranularity"
						}
					},
					{
						"in": "query",
						"name": "currency",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentCurrency"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "course",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payments/summary": {
			"get": {
				"operationId": "GetPaymentsSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentsSummary"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "currency",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentCurrency"
						}
					},
					{
						"in": "query",
						"name": "course",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payments/charts/status": {
			"get": {
				"operationId": "GetPaymentsStatusChart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/StatusChartItem"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Payments Charts"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "currency",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentCurrency"
						}
					},
					{
						"in": "query",
						"name": "course",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payments": {
			"get": {
				"operationId": "GetPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedPaymentsResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "currency",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentCurrency"
						}
					},
					{
						"in": "query",
						"name": "course",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentSortBy"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortOrder"
						}
					}
				]
			}
		},
		"/payments/export.csv": {
			"get": {
				"operationId": "ExportPaymentsCsv",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"text/csv": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentStatus"
						}
					},
					{
						"in": "query",
						"name": "currency",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/PaymentCurrency"
						}
					},
					{
						"in": "query",
						"name": "course",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}